Updated ./black-parrot-sim/docker-compose.yml to reflect new Dockerfile.ubuntu20.04 filename. Filename mismatch caused docker-compose build failure. After changing the filename in the 'dockerfile' field of docker-compose.yml, it now works.
I also updated the README because the docker-compose instructions as listed weren't working for me. I don't know if I was doing something wrong, but I listed the sequence of commands I used to succesfully run docker-compose (if it's helpful).

Lastly, I changed the line in Dockerfile.ubuntu20.04 that grabs the UID/GID of the user into a different argument (--user-group) because the original would not work properly for me... (I might've been doing something wrong though).

This is my first pull request, so I apologize if this is a little wonky. I hope these changes help!

Let me know if you have any questions,
Colin